media-viewer: add greace pencil layer creation per media file.
When media gets loaded a new greace pencil layer will be created in the corresponding greace pencil object. The layer is named after filepath. New layer will be set active and all other layers will be hidden. That way we can have annotation that are bound to one specific media file. If multiple videos are selected it will hide all layers for now.