media-viewer: add render review operator.
The render review operator renders out the current media file with annotations. This only works for the sequence editor as the image editor has no easy way to make an openGL render. The user can define an review output dir in the media view header, and the filepath will be computed based on the filename and the current time. That way the user can render out the same review and does not overwrite. This patch registers a property group on the window manager as well.