Bigger refactoring: Add PROD_CONTEXT and ASSET_CONTEXT.
This refactoring splits up BUILD_CONTEXT in PROD_CONTEXT, ASSET_CONTEXT, BUILD_CONTEXT.
The idea behind this is to have 3 contexts that we can update independent from each other. For example the
PROD_CONTEXT should be loaded on startup. It hold the production level defined task layers.

The ASSET_CONTEXT should be loaded on scene load. It holds the settings for the task layer per asset. Like which one is enabled and disabled. What is the asset collection in that scene, etc.

The BUILD_CONTEXT wil be initialized when User hits start publish, and holds the PROD and ASSET_CONTEXT as well as other information.