Use __setstate__, __getstate__ functions to modify pickle behavior of Context Classes.
Some objects cannot be pickled, like the blender_context or a collection, python module etc. By using the ddunder functions we can store these objects as a tmp data (like a string) and restore them properly on unpickeling. That way we can for example restore the asset collection and blender context easily.