AssetImporter: Add logic to detect base.
The base gets detected by checking if eneabled asset task layers have the lowest production order. If so the asset task will be used as a base and gets duplicated. If not the asset publish will be used as a base. If the asset publish is the base we use a trick and save the blend file in a tmp location and import the sdame publish collection again from it. That way we avoid annyoing .001 suffixes.