Add TransferSettings Concept.
Transfer settings are certain options Users might want to have to modify how data is transferred. Users can now define a  TransferSettings class in the task_layers.py module which is a regular Blender PropertyGroup. This PropertyGroup will be registered on scene level and can then be easily queried in the transfer data function of the TaskLayer. That way Users can provide themselves options to use in their code. This options are also displayed in the Blender AssetPipeline Panel automatically.