Add support for reading xml files and converting them to their respective MetadataClass. At pull support adjusting TaskLayer source path and updated_at key.
Metadata Classes can also be generated from ElementClasses. This conversion is happening in the .from_element() function.
The idea is that the code rest of the code only works with dataclasses and adjusts their data. That results in this logic.
A: Saving Metadata to file:
    -> MetadataClass -> ElementClass -> XML File on Disk
B: Loading Metadata from file:
    -> XML File on Disk -> ElementClass -> MetadataClass.