Add TaskLayerLockPlan concept.
This patch adds a TaskLayerLockPlan class that can save for an asset publish what TaskLayers need to be locked. This is used to handle the logic of locking previous TaskLayers of previous asset publishes when creating a new version. This information is also saved on scene level so it can be displayed in the UI after starting a new publish.