Add asset_updater module.
This patch adds some basics for the asset updater. The collect_assets operator scans the scene for imported asset collections. All this information is stored on a scene property. For each loaded asset we also load all available publishes and display them in an enum property in the same list row.