SVN: File status is now kept up to date entirely via background processes. This solution requires way less guesswork and hacking, as each file entry now keeps track of its working copy status separately from its remote repo status, which lets us do all the fancy things we were doing before in a way more reliable and less hacky way. Yay!
This can be a bit slow though (because SVN server is slow to respond), so operators need some manual file status setting for faster user feedback. This will always be overwritten a few seconds later by the automatic status, but hopefully that should result in no change, as long as we set the new statuses correctly.