SVN: Better prediction system to avoid flicker
Previously, a single flag was used to completely ignore the next single
file status update. Now, each file entry has a status_predicted_flag,
which will lets updates be ignored on files being updated or committed,
until the process is finished.
Also, for local svn operations that are instantaneous, we ignore a
single status update like before, but on a per-file basis,
so an svn status update is never fully discarded.