AssetPipeline: Iterate on UX based on feedback
Simon pointed out to me that allowing users to specify a single
task layer on a per-file basis is not necessarily the way to go,
since it would be nice to allow a file to own several task layers.

Still, the previous checkbox overload can be remedied by moving the
checkbox list to a separate panel, and letting the Pull operator show
the list of pull checkboxes on a pop-up, pre-configured
based on the permanent checkboxes.

The collection hierarchy transfer now relies on the Task Layer's suffix
being specified in each TaskLayer class.