GNSK: New Append/Link behaviour
Instead of always just linking the GeoNodes from the add-on's folder
(which is really bad for production),
We append by default, and allow linking from a custom location.
I also added some un-intrusive hard-coded heuristic, such
that if the BlenderSVN add-on is installed, and the current
file is in the Pet Projects production SVN,
get the file from the production's folder.