SVN: Massive UI drawing performance boost
It was taking a crazy amount of time to determine whether the SVN log
should even be drawn, and also determining whether each file should
be filtered or not.

Now instead of doing this at drawing time, we store a flag on each
file, and update that whenever the search filter changes,
and after status updates and operations.

Should be safe...? Maybe?

A similar change for SVN log entries might follow,
but the slow-down from that seems nowhere near as bad.