SVN: Optimize log filtering
Same method as previous commit:
Whenever the active file changes, we set flags on each log file,
indicating whether that file is affected.

This doesn't optimize text-based search filtering, only filtering by
the active file.

Either way, both of these types of filtering are already very fast,
so the speed-up is not as dramatic as the previous improvement.

And for that reason, I probably won't bother also optimizing the
text-based filtering.