Blender Crawl: During Purge Remove Fake User for Linked Data-Blocks (#105)
...and collections.

I believe this should be built-in Blender behaviour, but it's already on the [devs' radar](https://projects.blender.org/blender/blender/issues/106321), and in the meantime we have a production that will need to be cleaned up soon.

To explain a bit better, here's what can happen sometimes:
- You have an ID with a fake user in an asset (usually a collection)
- You link the asset to a shot
- You change your mind, and don't want the asset in the shot anymore; You might Un-link it from the Scene instead of "Delete Hierarchy".
- Now its collections have no real users left
- You run Purge and expect it to go away from the file
- It does not, because it has a fake user flag.

Even though obviously that fake user flag's intention was to keep the ID in the **original** file, not when it gets linked in somewhere else. Not to mention that people don't realize that Collections have the fake user flag set to True by default, since it's not visible *anywhere*.

Tested on Pets shots 50_70, 50_80, 50_90, 50_100.

Reviewed-on: https://projects.blender.org/studio/blender-studio-pipeline/pulls/105