Docs: Project Tools Update (#156)
This PR is to focus the studio pipeline documentation on the [project tools](https://projects.blender.org/studio/blender-studio-pipeline/src/branch/main/scripts/project-tools) scripts that are used for deployment and management of a project, instead of Gentoo. Gentoo documentation remains in a new section, while the TD / User guides are now focused on project tools.

## What's Changed?
 - Quick Setup Guide is now under TD Guide section with the name Project Tools Setup
 - Quick Start Guide is now under User Guide with the name Project Usage
 - Gentoo related articles have moved to their own section
 - Shared Folder Layout has been added to naming conventions

Reviewed-on: https://projects.blender.org/studio/blender-studio-pipeline/pulls/156