Asset Pipeline: Use original source data for UV seam transfer (#212)
The data transfer operator in Blender transfers between the original data of the target and the evaluated data of the source. This is an issue when there are modifiers that drastically alter the geometry, since they don't match.

With this PR I'm creating a proxy source for the transfer that uses no modifiers to basically transfer from and to original data.

Co-authored-by: Simon Thommes <simon.thommes@gmx.de>
Reviewed-on: https://projects.blender.org/studio/blender-studio-pipeline/pulls/212