Lighting Overrider: Move away from linking execution script (#237)
This PR gets rid of the concept of linking the execution script entirely, which previously was hard-coded to a certain file-path.
Now the addon will update the local execution script on every save and ensure that it is available.

The downside of this is that now, when changes to the execution script need to be made, every file that this change needs to propagate to will need to be updated as well.

Co-authored-by: Simon Thommes <simon.thommes@gmx.de>
Reviewed-on: https://projects.blender.org/studio/blender-studio-pipeline/pulls/237