Blender Kitsu: Add Operator to Import Playblasts into Edit (#274)
When adding the Playblast Operator I felt it was similar to the Import Image Sequence Operator so I made them both behave in a similar fashion, and added them to a new box, under the Shot Tools panel, which contains all the operators related to Metadata strips. Because we are using the Metadata strips (which have o task type) we must add the task type to the pop-up of both operators.

## What's Changed?
 - Add Operator to Import Playblasts from selected Metadata Strip
 - Update Import Image Sequence Operator to import from selected Metadata Strip (adds task type dropdown)
 - Add Media Panel to Shot Tools for Importing Media from a selected Metadata strip
 - Update Project Usage Guide with new Operators
 - Update Blender Kitsu README with new Operators

## New Panel
![image](/attachments/b7dd14bf-fa7b-4190-b0f2-c92c43c0a0a3)

## New Import Playblast Operator Pop-Up
![image](/attachments/5167e7f9-b442-46e5-b93e-d8e186d3d92d)
## Updated Import Image Sequence Pop-Up
![image](/attachments/a51489ef-db99-45a8-ae35-af96a58f5468)

Reviewed-on: https://projects.blender.org/studio/blender-studio-pipeline/pulls/274