fix: Make shot builder more scriptable (#292)
When executing the `KITSU_OT_build_new_shot` operator programmatically from a Python script, it's currently not possible to set `active_project`, which creates errors later in the `execute` method resulting in the scene only getting partially built. Not sure if there is an important reason why the global var was used directly for the project instead of using the cache to be consistent with how the other props are accessed in the `execute` method (e.g., shot, sequence), but this change doesn't seem to introduce any regressions from my tests. With this fix in place, a script like the following works:

```
import bpy

def generate_shot(production, project_id, episode, sequence, shot):
  kitsu_props = bpy.context.scene.kitsu
  kitsu_props.project_active_id = project_id
  kitsu_props.category = "SHOT"
  kitsu_props.episode_active_name = episode
  kitsu_props.sequence_active_name = sequence
  kitsu_props.shot_active_name = shot
  kitsu_props.task_type_active_name = "Animation"

  bpy.ops.kitsu.build_new_shot(production_name=production, save_file=True)
```

Reviewed-on: https://projects.blender.org/studio/blender-studio-pipeline/pulls/292