Blender Media Viewer: Add UI Scale Option (#294)
Closes: https://projects.blender.org/studio/blender-studio-pipeline/issues/281

Blender Media Viewer scales the view to 1.5X as opposed to the default 1X, we don't want to change that because in most cases this is the best UI scale, but it may be good to give the user ability to change this, espcially when working on small screens or tablets for example.

## What's Changed
 - Adds UI Scale Option to Settings

## Issues with this Fix
The UI Scale doesn't behave well when scaling elements, the menu it lives in needs to re-draw (only occurs after closing an opening the menu) not sure if this really acceptable as a fix because it makes the UI so ugly while changing scales. Aside from exposing the Preferences menu (which is not a good idea for a "Application Template") I don't see any solution to this issue

![expose_ui_scale](/attachments/c2d454d4-aa4f-45c1-9034-2c63ab66b6e2)

Reviewed-on: https://projects.blender.org/studio/blender-studio-pipeline/pulls/294
Reviewed-by: Sebastian Parborg <zeddb@noreply.localhost>