Update CloudRig docs & show in the sidebar
Added a whole new Workflow Enahncements page, and a bunch more videos, since videos actually work on this platform!

Still more work to be done though; The Introduction page is missing the contents of the repo's ReadMe basically. Also, a link to the repo itself will come, once it moves.