Render Review: Change heuristic to determine intended length from longest to latest (#309)
This is still not ideal, since the latest render can be an incomplete one.
The problem with the previous method was that this feature would not work at all when a shot was shortened, since the longer version was always interpreted as the correct length.

While it is still flawed until we use the actual intended frame length to compare, now at least it is explicit.

Reviewed-on: https://projects.blender.org/studio/blender-studio-pipeline/pulls/309