Kitsu: Fix bulk shot initialization
The manual_shot_name prop was being evaluated exclusively (ignoring
shot_name, which is set on shot init). This made it impossible to
submit newly initialized shots unless a manual_shot_name was set.

Now both props are taken into account, and if manual_shot_name is not
set, we show show_name as placeholder.