Fix #385: Contactsheet broken by API changes
In Blender 4.4, bpy.types.Sequence was renamed to bpy.types.Strip.
In the codebase, this type was only used for type annotations, so to preserve backwards compatibility and keep this quick, I just changed it to a string, "bpy.types.Strip". That can be turned back from a string to a real ref if we ever want to break pre-Blender 4.4 compatibility.