Brushstroke Tools: Link brushstrokes to same collections as surface object rather than the active collection (#390)
This changes it so that new brushstroke and flow objects are linked to the same collections as the original surface object they relate to.

Any linked collections are skipped. If none remains, the active collection is used as a fallback. If that is also linked, the scene collection is used to ensure the objects to be immediately editable and visible.

Reviewed-on: https://projects.blender.org/studio/blender-studio-tools/pulls/390