Asset P: Bump min. ver. to 5.0
This is due to non-backward-compatible PyAPI changes to how properties are accessed, which were made earlier today.

In other add-ons it was easier to add backwards compatibility code, but here I deemed it too much hassle, as there's quite a bit of code dealing with custom properties. A version from yesterday should still work going back to 4.1.