Asset P: Fix error in modifier transfer
Not sure why this started happening, but for object pointer values, it was failing to type-cast them.
Removing the type casting altogether resulted in errors trying to assign integers to booleans, which I also don't understand.
Temporary solution for now, just hard-coded some supported types that should be safe to type cast.