Project Tools: Ensure Stable download on non-main Blender Branches (#425)
When updating Blender, when a user has selected to be on a stable branch, we want to filter for a given risk ID. In most cases, users will want to be on the stable branch. In other cases, they may want to customize this. Add the ability to filter by risk_id when not set to the main branch and set default value to "stable".

Reviewed-on: https://projects.blender.org/studio/blender-studio-tools/pulls/425