Project Tools: symlink-patch-windows (#460)
Using windows 11, I have created a project path as follows:
  %HOMEPATH%/capstone

Within that folder i have used the windows mklink command to create 2 symlinks to smb shares:
  svn -> \\smb_server\blender_svn
  shared -> \\smb_server\blender_shared

When deployment_assistant is run, the PATH_BASE and PATH_CUSTOM_SPLASH variables are being set to "\\server\blender_svn" rather than the expected "%HOMEPATH%/capstone".  Because of this, none of the other PATH variables that are based off of these are set properly and the application fails due to inability to resolve paths properly.

This is caused by the use of resolve() in the calls to "Path(__file__).resolve().parent.parent.parent".  Removing the call to resolve() fixes the issue and the application is able to properly execute.

Reviewed-on: https://projects.blender.org/studio/blender-studio-tools/pulls/460