Project Tools: Determine `project_root_dir` during Deployment Assistant runtime (#461)
This change modifies the setting of the "project_root_dir" to be determined at runtime via the python Path object rather than being loaded from the "project_config.json" file.  This allows the pathing to be properly set regardless of the operating system on the artist system.  In addition, the "project_root_dir" is no loger being stored in the "project_config.json" file as it is no longer needed.

Reviewed-on: https://projects.blender.org/studio/blender-studio-tools/pulls/461