Blender Kitsu: Improve Fallbacks when getting Kitsu Task Types (#479)
Ran into this issue on a production I am supporting.

Found that Shot Builder will fallback to shortname if not found in task mapping, for a shot and assets will fall back to the actual name (not basename). That is an inconsistency that should be resolved. Even if resolved context doesn't support this, it only goes off of mapping. The inconsistencies with fetching information like from `filepath->context` and `context->filepath` are something to address in a future refactor. But to fix the context side of things I have come up with a simple fix.

1. Check if we find our key in mapping
2. Check if we find a task type by name (Assets outside of mapping)
3. Check if we find a task type by shortname (Shots outside of mapping)

Then finally return None. Allow the context detect to return whatever error message, instead of also reporting this in a separate method. Reduces error messages but increases clarity.

This can be packported to 4.5

Reviewed-on: https://projects.blender.org/studio/blender-studio-tools/pulls/479